Create InsertPosition.cpp
Input Format 
You have to complete the Node* Insert(Node* head, int data, int position) method which takes three arguments - the head of the linked list, the integer to insert and the position at which the integer must be inserted. You should NOT read any input from stdin/console. position will always be between 0 and the number of the elements in the list (inclusive).

Output Format 
Insert the new node at the desired position and return the head of the updated linked list. Do NOT print anything to stdout/console.