Create DeleteNode.cpp
Input Format 
You have to complete the Node* Delete(Node* head, int position) method which takes two arguments - the head of the linked list and the position of the node to delete. You should NOT read any input from stdin/console. position will always be at least 0 and less than the number of the elements in the list.

Output Format 
Delete the node at the given position and return the head of the updated linked list. Do NOT print anything to stdout/console.